# [LinearSolver.Direct] Parallelization of H A^-1 H^T in SparseLDLSolver