--- a/Sofa/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLDLSolver.h
+++ b/Sofa/Component/LinearSolver/Direct/src/sofa/component/linearsolver/direct/SparseLDLSolver.h
@@ -56,6 +56,7 @@ public :
     typedef typename Inherit::JMatrixType JMatrixType;
     typedef SparseLDLImplInvertData<type::vector<int>, type::vector<Real> > InvertData;
 
+    void init() override;
     void parse( sofa::core::objectmodel::BaseObjectDescription* arg ) override;
     void solve (Matrix& M, Vector& x, Vector& b) override;
     void invert(Matrix& M) override;
@@ -72,6 +73,8 @@ public :
     SOFA_ATTRIBUTE_DISABLED__SPARSELDLSOLVER_MATRIXEXPORT
     DeprecatedAndRemoved d_precision;
 
+    Data<bool> d_parallelInverseProduct;
+
     MatrixInvertData * createInvertData() override {
         return new InvertData();
     }
@@ -79,7 +82,7 @@ public :
 protected :
     SparseLDLSolver();
 
-    type::vector<int> Jlocal2global;
+    type::vector<sofa::SignedIndex> Jlocal2global;
     sofa::linearalgebra::FullMatrix<Real> JLinvDinv, JLinv;
     sofa::linearalgebra::CompressedRowSparseMatrix<Real> Mfiltered;