--- a/examples/RegressionStateScenes.regression-tests
+++ b/examples/RegressionStateScenes.regression-tests
@@ -16,7 +16,7 @@ Demos/chainHybridNoGroup.scn 100 1e-2 1 1
 Demos/collisionMultiple.scn 100 1e-4 1 1
 Demos/liver.scn 100 1e-4 1 1
 Demos/simpleBoundaryConditions.scn 100 1e-4 1 1
-Demos/fallingBeamLagrangianCollision.scn 46 1e-4 1 1
+Demos/fallingBeamLagrangianCollision.scn 46 1e-8 1 1
 
 
 ### Component scenes ###