The parallelized function is `SparseLDLSolver::doAddJMInvJtLocal`.

# Benchmarks

I rely on the scene provided by @bakpaul with light changes: https://gist.github.com/alxbilger/1b82b1488a2741be845d7d9da18c25e5. The SOFA timer to focus on is `Get Compliance` as it is the deepest timer in the hierarchy containing the `doAddJMInvJtLocal` function.

- Master branch (commit df49f3a754264489b14db69b38862a5b45d57168)

```
100 iterations done in 12.9731 s ( 7.70827 FPS).
```

```
 LEVEL   START    NUM      MIN     MAX   MEAN     DEV    TOTAL  PERCENT ID
   5      15.41    1       0.01  121.72   98.03   20.57   98.03   77.48 .....Get Compliance
```

- This PR (without parallelization)

The benchmark should be similar to the master branch

```
100 iterations done in 13.1572 s ( 7.6004 FPS).
```

```
 LEVEL   START    NUM      MIN     MAX   MEAN     DEV    TOTAL  PERCENT ID
   5      16.31    1       0.01  134.76   98.34   22.21   98.34   76.54 .....Get Compliance
```

- This PR (with parallelization)

```
100 iterations done in 5.94345 s ( 16.8253 FPS).
```

```
 LEVEL   START    NUM      MIN     MAX   MEAN     DEV    TOTAL  PERCENT ID
   5      15.76    1       0.01   30.62   26.85    5.55   26.85   47.91 .....Get Compliance
```

On my machine there is a nice speedup of x3.66 on the `Get Compliance` timer, and x2.18 in terms of global FPS.

# Additional Benchmarks

I wanted to make sure that the usual solve of SparseLDLSolver did not regress in terms of performances. I benchmarked `examples/Component/LinearSolver/Direct/FEMBAR_SparseLDLSolver.scn` on 10000 steps.

- Master branch (commit df49f3a754264489b14db69b38862a5b45d57168)

```
10000 iterations done in 28.2657 s ( 353.785 FPS).
```

- This PR

```
10000 iterations done in 28.3223 s ( 353.078 FPS).
```

______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).