@fredroy 2023-07-11
I did some benches on TorusFall.scn🤐
master 
`5000 iterations done in 41.7616 s ( 119.727 FPS)`

pr       
 `5000 iterations done in 36.09 s ( 138.543 FPS)`