@alxbilger 2023-07-13
@bakpaul thanks for the benchmark. Yes, I would like your scene to investigate both the behavior and the performances (I am a bit disappointed so far).