@bakpaul 2023-07-13
Here it is ! 
[625-Falling-Lagrangian.zip](https://github.com/sofa-framework/sofa/files/12036504/625-Falling-Lagrangian.zip)