@bakpaul 2023-07-18
Nice job, it is indeed faster on my computer too. 
For the artifact, it still appears at the top of the beam on the contact, at step 46 (see image) event with your modified scene (I saw only the ShewchukPCGLinearSolver that was different).  

![image](https://github.com/sofa-framework/sofa/assets/30337881/45010585-c1e3-4f89-9e9b-f201bd700a29)

Would be great to know what are the difference between our compilation configurations.