@fredroy 2023-07-20
> @bakpaul Strange, I checked again (with and without parallelization) and I don't have the artefact you have. `ShewchukPCGLinearSolver` is not a difference because you use an alias of it (`PCGLinearSolver`). In the benchmark I translated the beam to measure mainly the contacts and skip the falling part. But I still don't see the artefact when I remove the translation. @fredroy @hugtalbot would you like to test this branch with the scene provided by @bakpaul ?

I tested the scene of @bakpaul and I get the same artefact on ubuntu 23.04/gcc12.2