@alxbilger 2023-07-20
@fredroy Thanks for trying.
So the regression depends on the OS. Great!
I added a regression test for this scene in https://github.com/sofa-framework/sofa/pull/4055