@alxbilger 2023-07-27
Some notes:

- It was extremely hard to understand why L^T was used in the first step of the solve, and not L. I hope it's clearer now with the comment I added. Anyway, it requires to have the knowledge of the CSR and CSC formats, and how they relates to the transpose.
- I would like to use C++20 `std::span` in some functions such as `solveLowerUnitriangularSystemCSR`.
- Regression tests threshold are usually set to 1e-4. IMO it's not enough to capture errors happening on a single DOF because it's an average